Rename all occurences of codis with disco
