+++ .github/workflows/workflow.yml
@@ -27,6 +27,6 @@ jobs:
     - name: Run ruff
       uses: chartboost/ruff-action@v1
     - name: Run black
-      run: black --check --verbose --diff ./codis
+      run: black --check --verbose --diff ./disco
     - name: Run pytest
       run: python -m pytest ./tests
