--- codis/models/__init__.py
@@ -1,4 +1,0 @@
-"""Models and model blocks."""
-from codis.models.autoencoder import AutoEncoder
-from codis.models.beta_vae import BaseVAE, BetaVAE
-from codis.models.mlp import MLP
