+++ configs/model/vae.yaml
@@ -1,6 +1,6 @@
-_target_: codis.lightning.modules.SupervisedVAE
+_target_: disco.lightning.modules.SupervisedVAE
 vae:
-  _target_: codis.lightning.modules.LightningBetaVAE
+  _target_: disco.lightning.modules.LightningBetaVAE
   img_size: ${dataset.img_size}
   latent_dim: 10
   beta: 1.0
