+++ disco/data/__init__.py
@@ -1,5 +1,5 @@
 """Datasets and data utils."""
-from codis.data.infinite_dsprites import (
+from disco.data.infinite_dsprites import (
     InfiniteDSprites,
     ContinualDSpritesMap,
     RandomDSprites,
@@ -8,7 +8,7 @@ from codis.data.infinite_dsprites import (
     InfiniteDSpritesTriplets,
     Latents,
 )
-from codis.data.continual_benchmark import (
+from disco.data.continual_benchmark import (
     ContinualBenchmark,
     ContinualBenchmarkRehearsal,
 )
