+++ disco/data/infinite_dsprites.py
