+++ disco/models/__init__.py
@@ -1,0 +1,4 @@
+"""Models and model blocks."""
+from disco.models.autoencoder import AutoEncoder
+from disco.models.beta_vae import BaseVAE, BetaVAE
+from disco.models.mlp import MLP
