+++ disco/models/base_vae.py
