+++ scripts/visualize_joint_training.py
@@ -62,8 +62,8 @@ def _main():
     parser.add_argument(
         "--wandb_entity",
         type=str,
-        help="Wandb entity and project name, e.g. codis/codis",
-        default="codis/codis",
+        help="Wandb entity and project name, e.g. disco/disco",
+        default="disco/disco",
     )
     parser.add_argument("--wandb_group", type=str, help="Wandb group name")
     parser.add_argument(
